updated logic

+++ NumberClassificationApi/Models/NumberProperties.cs
@@ -2,7 +2,7 @@
 {
     public class NumberProperties
     {
-        public int number { get; set; }
+        public double number { get; set; }
         public bool is_prime { get; set; }
         public bool is_perfect { get; set; }
         public List<string> properties { get; set; } = new List<string>();
